requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
google-api-python-client==2.103.0